golang lcs serializer & deserializer implementation round one (#46)
* golang lcs serializer & deserializer implementation, map is not supported yet
* update golang generation test to import module instead of customizing GOPATH
* build env update: install go 1.13 version
* move runtime/golang/src to runtime/golang
* fix SortMapEntries, GetBufferOffset, SerializeLen interfaces

+++ serde-generate/runtime/golang/serde/interfaces.go
@@ -38,15 +38,15 @@ type Serializer interface {
 
 	SerializeI128(value Int128) error
 
-	SerializeLen(value int) error
+	SerializeLen(value uint64) error
 
 	SerializeVariantIndex(value uint32) error
 
 	SerializeOptionTag(value bool) error
 
-	GetBufferOffset() int
+	GetBufferOffset() uint64
 
-	SortMapEntries(offsets []int)
+	SortMapEntries(offsets []uint64)
 
 	GetBytes() []byte
 }
@@ -86,9 +86,9 @@ type Deserializer interface {
 
 	DeserializeI128() (Int128, error)
 
-	DeserializeLen() (int, error)
+	DeserializeLen() (uint64, error)
 
-	DeserializeVariantIndex() (uint64, error)
+	DeserializeVariantIndex() (uint32, error)
 
 	DeserializeOptionTag() (bool, error)
